config/settings.py
"""
config/settings.py
Central configuration for the Hospital Analytics Pipeline.
All paths, credentials (via env), and tunable parameters live here.
"""

import os
from dataclasses import dataclass, field
from pathlib import Path

try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

BASE_DIR = Path(__file__).resolve().parent.parent
DATA_DIR = BASE_DIR / "data"


@dataclass
class DatabaseConfig:
    host: str = os.getenv("DB_HOST", "localhost")
    port: int = int(os.getenv("DB_PORT", 5432))
    name: str = os.getenv("DB_NAME", "hospital_db")
    user: str = os.getenv("DB_USER", "analytics_user")
    password: str = os.getenv("DB_PASSWORD")
    
    @property
    def url(self) -> str:
        return (
            f"postgresql://{self.user}:{self.password}"
            f"@{self.host}:{self.port}/{self.name}"
        )
