requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0       # Parquet support

# Scheduling
APScheduler>=3.10.0

# Environment / config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: replace SQLite with a real warehouse
psycopg2-binary>=2.9.0      # PostgreSQL
# snowflake-connector-python  # Snowflake
# google-cloud-bigquery        # BigQuery
